refactor(lexer): 将Lexer.h中的start和current字段类型改为size_t
refactor(interpreter): 为Environment及其子类添加虚析构函数

fix(interpreter): 在Constructor.cpp中抑制未使用参数警告

build: 更新CMakeLists.txt配置，改进构建系统

style(lexer): 调整Lexer.cpp代码格式和include顺序

refactor(interpreter): 重构Property.cpp代码结构并抑制未使用变量警告

+++ src/lexer/Lexer.h
@@ -29,8 +29,8 @@ private:
   bool isUnicodeIdentifierContinue(char c);
 
   std::string source;
-  int start = 0;
-  int current = 0;
+  size_t start = 0;
+  size_t current = 0;
   int line = 1;
   bool inInterpolation = false;
   bool inStringLiteral = false; // Track if we're inside a string literal
